pyproject.toml
[project]
name = "resolve-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Kevyn Menezes",email = "shamior_@hotmail.com"}
]
requires-python = ">=3.14,<4.0"
dependencies = [
    "fastapi[standard] (>=0.133.1,<0.134.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pycountry (>=26.2.16,<27.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
]

[tool.poetry]
packages = [{include = "app"}]

[tool.ruff]
builtins = ["_"]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = [
    "I",        # Ordena as importações direitinho
    "F",        # Regras do Pyflakes
    "E",        # Erros do PyCodeStyle
    "W",        # Avisos do PyCodeStyle
    "PL",       # pylint
    "PT",       # pytest
    "FAST",     # fastapi
    "COM",      # impõe regras de vírgula final
    "N",        # impõe convenções de nomenclatura, e.g., ClassName vs function_name
    "C4",       # Pega o uso incorreto de comprehensions, dict, list, etc
    "ISC",      # Bom uso de concatenação de strings
    "SIM",      # Regras comuns de simplificação
    "TID",      # Algumas boas práticas de importação
]
ignore = [
    "PLR0913",  # Too many arguments for a function
    "PLR0917",  # Too many positional arguments
]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"


[tool.coverage.run]
omit = [
    "app/infra/database/seed/seed.py",
]


[tool.taskipy.tasks]
lint = "ruff check"

pre_format = "ruff check --fix"
format = "ruff format"

run = "fastapi dev app/main.py"

pre_test = "task lint"
test = "pytest -s -x --cov=app -vv"
post_test = "coverage html"

pre_ptest = "task pre_test"
ptest = "pytest -s -vv"

gen_migration = "alembic revision --autogenerate -m"
migrate = "alembic upgrade head"

drop_migrations = "alembic downgrade base"
post_drop_migrations = "task migrate"

pre_seed = "task drop_migrations"
seed = "python -m app.infra.database.seed.seed"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.5,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
